+++ addon/components/object-list-view.js
@@ -1585,6 +1585,8 @@ export default FlexberryBaseComponent.extend(
         return ['eq', 'neq', 'le', 'ge'];
 
       case 'string':
+        return ['eq', 'neq', 'like'];
+
       case 'boolean':
         return ['eq', 'neq'];
 
@@ -1724,7 +1726,7 @@ export default FlexberryBaseComponent.extend(
           let filters = {};
           let hasFilters = false;
           this.get('columns').forEach((column) => {
-            if (column.filter.condition) {
+            if (column.filter.condition || column.filter.pattern) {
               hasFilters = true;
               filters[column.filter.name] = column.filter;
             }
@@ -1733,7 +1735,11 @@ export default FlexberryBaseComponent.extend(
           if (hasFilters) {
             this.sendAction('applyFilters', filters);
           } else {
-            this.get('currentController').send('refreshList');
+            if (this.get('currentController.filters')) {
+              this.get('currentController').send('resetFilters');
+            } else {
+              this.get('currentController').send('refreshList');
+            }
           }
         } else {
           this.get('currentController').send('refreshList');