+++ addon/mixins/flexberry-objectlistview-on-edit-form-controller.js
@@ -6,7 +6,7 @@ import Ember from 'ember';
 import { Query } from 'ember-flexberry-data';
 import deserializeSortingParam from '../utils/deserialize-sorting-param';
 
-const { Condition, SimplePredicate, ComplexPredicate, DatePredicate, Builder } = Query;
+const { Condition, SimplePredicate, StringPredicate, ComplexPredicate, DatePredicate, Builder } = Query;
 
 /**
   Mixin for edit-form-controller for ObjectListView support.
@@ -211,6 +211,9 @@ export default Ember.Mixin.create({
     if (filter.pattern && filter.condition) {
       switch (filter.type) {
         case 'string':
+          return filter.condition === 'like' ?
+            new StringPredicate(filter.name).contains(filter.pattern) :
+            new SimplePredicate(filter.name, filter.condition, filter.pattern);
         case 'boolean':
           return new SimplePredicate(filter.name, filter.condition, filter.pattern);
         case 'number':
@@ -220,6 +223,11 @@ export default Ember.Mixin.create({
 
         default:
           return null;
+      }
+    } else {
+      if (!filter.condition && filter.type === 'string') {
+        Ember.set(filter, 'condition', 'like');
+        return new StringPredicate(filter.name).contains(filter.pattern);
       }
     }